equitas_ai/requirements.txt
fastapi
uvicorn
python-multipart
pandas
scikit-learn
shap
google-generativeai
langgraph
reportlab
python-dotenv